% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{download_heatmap}
\alias{download_heatmap}
\title{Show recorded heatmap}
\usage{
download_heatmap(
  path = "www/heatmap-data.json",
  filename = "heatmap.png",
  session = shiny::getDefaultReactiveDomain(),
  clean = TRUE,
  options = NULL
)
}
\arguments{
\item{path}{Previously saved heatmap data for persistence.}

\item{filename}{Screenshot file name.}

\item{session}{Shiny session object.}

\item{clean}{Whether to clean the heatmap. This allows to click
on the UI again and generate a new heatmap. Default to TRUE.}

\item{options}{Slot for heatmap options. Expect a (nested) list.
See \url{https://www.patrick-wied.at/static/heatmapjs/docs.html#heatmap-configure}.}
}
\description{
Show and download \link{record_heatmap} data.
}
