% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{get_heatmap}
\alias{get_heatmap}
\title{Get heatmap from running Shiny app.}
\usage{
get_heatmap(url, options = NULL)
}
\arguments{
\item{url}{Shiny app url. Can be localhost or
server deployed app.}

\item{options}{Extra options to pass to shinytest2.}
}
\value{
Run the Shiny app with heatmap display.
}
\description{
This is the default function to use to collect
heatmap data.
}
