% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{get_heatmap_records}
\alias{get_heatmap_records}
\title{Get heatmap file records}
\usage{
get_heatmap_records(path)
}
\arguments{
\item{path}{Recordings location}
}
\value{
A vector containing full path to recordings.
}
\description{
List all heatmap recordings
}
