% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_container}
\alias{heatmap_container}
\title{Heatmap container for shiny app}
\usage{
heatmap_container(...)
}
\arguments{
\item{...}{Shiny UI code.}
}
\description{
Include necessary dependencies.
}
