% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{init_heatmap_storage}
\alias{init_heatmap_storage}
\title{Initialize heatmap data temporary storage.}
\usage{
init_heatmap_storage(session, target)
}
\arguments{
\item{session}{Shiny session object. Useful to store heatmap data.}

\item{target}{Container selector hosting the heatmap canvas.
Default to Shiny fluidPage container. Be careful to change it if
you use another template.}
}
\value{
Side effect to create a list in
session$userData and store the heatmap data.
}
\description{
Initialize heatmap data temporary storage.
}
\keyword{internal}
