% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{process_heatmap}
\alias{process_heatmap}
\title{Either record or show heatmap data}
\usage{
process_heatmap(...)
}
\arguments{
\item{...}{Pass in parameters for \link{record_heatmap} and
\link{download_heatmap}.}
}
\description{
By default, \link{process_heatmap} records the heatmap.
However, if you browse to the app url and add \code{?get_heatmap} query string,
the function will run in display mode to see the heatmap data.
}
\note{
Specify \code{options("shinyHeatmap.debug" = TRUE)} to
get useful debugging messages, particularly when setting up
your heatmap.
}
