% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_heatmap_records}
\alias{read_heatmap_records}
\title{Read and aggregate heatmap records}
\usage{
read_heatmap_records(records, viewport_dims)
}
\arguments{
\item{records}{Returned by \link{get_heatmap_records}.}

\item{viewport_dims}{Obtained from JS and stored in session$input$viewport_dims.}
}
\value{
A list with aggregated data.
}
\description{
Scale coordinates to make sure they render the same on all screens.
}
\keyword{internal}
