% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{record_heatmap}
\alias{record_heatmap}
\title{Record heatmap data to a file}
\usage{
record_heatmap(
  path = "www/heatmap-data.json",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{path}{Previously saved heatmap data for persistence.}

\item{session}{Shiny session object. Useful to store heatmap data.}
}
\description{
On the JS side, Shiny creates a input$heatmap_data
containing the current click data to be recorded
to the file
}
