% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{take_heatmap_screenshot}
\alias{take_heatmap_screenshot}
\title{Take heatmap screenshot}
\usage{
take_heatmap_screenshot(filename, target)
}
\arguments{
\item{filename}{Where to save the screenshot.}

\item{target}{Element to capture within the page. CSS
selector ...}
}
\description{
Take heatmap screenshot
}
\keyword{internal}
