% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{validate_heatmap_trigger}
\alias{validate_heatmap_trigger}
\title{Safety checks}
\usage{
validate_heatmap_trigger(trigger)
}
\arguments{
\item{trigger}{Reactive trigger to initialized the heatmap
recording. This is useful if your app contains tabs or navbar.}
}
\value{
Stop app if conditions are not met.
}
\description{
Safety checks
}
\keyword{internal}
