% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xia.R
\name{xia}
\alias{xia}
\title{Retrieve target LC/MS data published by Xia et al, 2022}
\source{
[Xia et al, Molecular neurodegeneration, 2022](https://molecularneurodegeneration.biomedcentral.com/articles/10.1186/s13024-022-00547-7)
}
\usage{
xia(
  contrast = c("1. APP-SAA heterozygo KI vs WT", "2. APP-SAA homozygous KI vs WT",
    "3. Hom_vs_Het")
)
}
\arguments{
\item{contrast}{Scalar character, the comparison of interest}
}
\value{
A list with four elements:
- abundance: a numeric feature x sample matrix with 151 rows and 18 columns
- stats: a `data.frame` with 145 rows and 6 columns
- sample_anno: a `data.frame` with 18 rows and 4 columns
- feature_anno: a `data.frame` with 151 rows and 0 columns
}
\description{
Retrieve target LC/MS data published by Xia et al, 2022
}
\examples{
# retrieve the four different objects
dl <- xia()
sapply(dl, dim)
sapply(dl, \(x) class(x)[1])
}
