% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{download_heatmap}
\alias{download_heatmap}
\title{Show recorded heatmap}
\usage{
download_heatmap(
  path = "www",
  filename = "heatmap.png",
  target = ".container-fluid",
  timeout = 10,
  session = shiny::getDefaultReactiveDomain(),
  options = NULL
)
}
\arguments{
\item{path}{Previously saved heatmap data for persistence.}

\item{filename}{Screenshot file name.}

\item{target}{Container selector hosting the heatmap canvas.
Default to Shiny fluidPage container. Be careful to change it if
you use another template.}

\item{timeout}{Necessary if the page needs time to load.
Expressed in milliseconds.}

\item{session}{Shiny session object.}

\item{options}{Slot for heatmap options. Expect a (nested) list.
See \url{https://www.patrick-wied.at/static/heatmapjs/docs.html#heatmap-configure}.}
}
\description{
Show and download \link{record_heatmap} data.
}
