% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{download_heatmap}
\alias{download_heatmap}
\title{Show recorder heatmap}
\usage{
download_heatmap(
  path = "www/heatmap-data.json",
  filename = "heatmap.png",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{path}{Previously saved heatmap data for persistence.}

\item{filename}{Screenshot file name.}

\item{session}{Shiny session object.}
}
\description{
Require to have \link{record_heatmap} data.
}
