% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{download_heatmap}
\alias{download_heatmap}
\title{Show recorded heatmap}
\usage{
download_heatmap(
  trigger = NULL,
  path = "www",
  filename = "heatmap.png",
  timeout = 10,
  show_ui = TRUE,
  options = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{trigger}{Reactive trigger to initialized the heatmap
recording. This is useful if your app contains tabs or navbar.}

\item{path}{Previously saved heatmap data for persistence.}

\item{filename}{Screenshot file name.}

\item{timeout}{Necessary if the page needs time to load.
Expressed in milliseconds.}

\item{show_ui}{Whether to show the download UI. Default to TRUE.}

\item{options}{Slot for heatmap options. Expect a (nested) list.
See \url{https://www.patrick-wied.at/static/heatmapjs/docs.html#heatmap-configure}.}

\item{session}{Shiny session object. Useful to store heatmap data.}
}
\description{
Show and download \link{record_heatmap} data.
}
