% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{process_heatmap}
\alias{process_heatmap}
\title{Either record or show heatmap depending on
configuration.}
\usage{
process_heatmap(...)
}
\arguments{
\item{...}{Pass in parameters for \link{record_heatmap} and
\link{download_heatmap}.}
}
\description{
By default, \link{process_heatmap} records the heatmap.
However, if you set \code{options("shinyheatmap.mode" = 'display')},
the function will run in display mode to see the results.
This is useful if you want to use shinytest2 to control
the app and display the results without affecting the deployed
production app.
}
