% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{record_heatmap}
\alias{record_heatmap}
\title{Record heatmap data to a file}
\usage{
record_heatmap(
  path = "www",
  target = ".container-fluid",
  type = c("click", "move"),
  timeout = 10,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{path}{Previously saved heatmap data for persistence.}

\item{target}{Container selector hosting the heatmap canvas.
Default to Shiny fluidPage container. Be careful to change it if
you use another template.}

\item{type}{Event type: click or mouse move. Default to click.}

\item{timeout}{Necessary if the page needs time to load.
Expressed in milliseconds.}

\item{session}{Shiny session object. Useful to store heatmap data.}
}
\description{
On the JS side, Shiny creates a input$heatmap_data
containing the current click data to be recorded
to the file
}
